pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Nikita <nixon1111110@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-environ = "^0.9.0"
psycopg2 = "^2.9.5"
Django = "4.0.1"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.0.0"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.24.2"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
